test(react-query): refactor test file structure 1 file by 1 api
